+++ pkg/kwp2000/kwp2000.go
@@ -957,7 +957,7 @@ func (t *Client) RequestUpload(ctx context.Context, address, length uint32) erro
 func (t *Client) sendLong(ctx context.Context, data []byte) error {
 	messages := t.splitRequest(data, true)
 	for i, msg := range messages {
-		//log.Println(msg.String())
+		log.Println(msg.String())
 		if i == len(messages)-1 {
 			//if err := t.c.Send(msg); err != nil {
 			//	return err